Fix int comparison with vector::size.
TBR=titzer@chromium.org

Review URL: https://codereview.chromium.org/503023002

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@23363 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/node.cc
@@ -18,7 +18,7 @@ void Node::CollectProjections(NodeVector* projections) {
     if ((*i)->opcode() != IrOpcode::kProjection) continue;
     int32_t index = OpParameter<int32_t>(*i);
     DCHECK_GE(index, 0);
-    DCHECK_LT(index, projections->size());
+    DCHECK_LT(index, static_cast<int32_t>(projections->size()));
     DCHECK_EQ(NULL, (*projections)[index]);
     (*projections)[index] = *i;
   }
